Use less memory for TriangleMesh and Triangle (usize -> u32).

+++ src/shapes/triangle.rs
@@ -23,7 +23,7 @@ use crate::core::transform::Transform;
 #[derive(Clone)]
 pub struct TriangleMesh {
     /// the total number of triangles in the mesh
-    pub n_triangles: usize,
+    pub n_triangles: u32,
     /// vector of vertex indices
     pub vertex_indices: Vec<usize>,
     /// the total number of vertices in the mesh
@@ -50,7 +50,7 @@ impl TriangleMesh {
         object_to_world: Transform,
         world_to_object: Transform,
         reverse_orientation: bool,
-        n_triangles: usize,
+        n_triangles: u32,
         vertex_indices: Vec<usize>,
         n_vertices: usize,
         p: Vec<Point3f>,
@@ -83,7 +83,7 @@ impl TriangleMesh {
 #[derive(Clone)]
 pub struct Triangle {
     mesh: Arc<TriangleMesh>,
-    pub id: usize,
+    pub id: u32,
     // inherited from class Shape (see shape.h)
     object_to_world: Transform,
     world_to_object: Transform,
@@ -98,7 +98,7 @@ impl Triangle {
         world_to_object: Transform,
         reverse_orientation: bool,
         mesh: Arc<TriangleMesh>,
-        tri_number: usize,
+        tri_number: u32,
     ) -> Self {
         Triangle {
             mesh,
@@ -119,17 +119,17 @@ impl Triangle {
             ]
         } else {
             [
-                self.mesh.uv[self.mesh.vertex_indices[self.id * 3 + 0]],
-                self.mesh.uv[self.mesh.vertex_indices[self.id * 3 + 1]],
-                self.mesh.uv[self.mesh.vertex_indices[self.id * 3 + 2]],
+                self.mesh.uv[self.mesh.vertex_indices[(self.id * 3) as usize + 0]],
+                self.mesh.uv[self.mesh.vertex_indices[(self.id * 3) as usize + 1]],
+                self.mesh.uv[self.mesh.vertex_indices[(self.id * 3) as usize + 2]],
             ]
         }
     }
     // Shape
     pub fn object_bound(&self) -> Bounds3f {
-        let p0: Point3f = self.mesh.p[self.mesh.vertex_indices[self.id * 3 + 0]];
-        let p1: Point3f = self.mesh.p[self.mesh.vertex_indices[self.id * 3 + 1]];
-        let p2: Point3f = self.mesh.p[self.mesh.vertex_indices[self.id * 3 + 2]];
+        let p0: Point3f = self.mesh.p[self.mesh.vertex_indices[(self.id * 3) as usize + 0]];
+        let p1: Point3f = self.mesh.p[self.mesh.vertex_indices[(self.id * 3) as usize + 1]];
+        let p2: Point3f = self.mesh.p[self.mesh.vertex_indices[(self.id * 3) as usize + 2]];
         bnd3_union_pnt3(
             &Bounds3f::new(
                 self.world_to_object.transform_point(&p0),
@@ -139,16 +139,16 @@ impl Triangle {
         )
     }
     pub fn world_bound(&self) -> Bounds3f {
-        let p0: Point3f = self.mesh.p[self.mesh.vertex_indices[self.id * 3 + 0]];
-        let p1: Point3f = self.mesh.p[self.mesh.vertex_indices[self.id * 3 + 1]];
-        let p2: Point3f = self.mesh.p[self.mesh.vertex_indices[self.id * 3 + 2]];
+        let p0: Point3f = self.mesh.p[self.mesh.vertex_indices[(self.id * 3) as usize + 0]];
+        let p1: Point3f = self.mesh.p[self.mesh.vertex_indices[(self.id * 3) as usize + 1]];
+        let p2: Point3f = self.mesh.p[self.mesh.vertex_indices[(self.id * 3) as usize + 2]];
         bnd3_union_pnt3(&Bounds3f::new(p0, p1), &p2)
     }
     pub fn intersect(&self, ray: &Ray) -> Option<(SurfaceInteraction, Float)> {
         // get triangle vertices in _p0_, _p1_, and _p2_
-        let p0: Point3f = self.mesh.p[self.mesh.vertex_indices[self.id * 3 + 0]];
-        let p1: Point3f = self.mesh.p[self.mesh.vertex_indices[self.id * 3 + 1]];
-        let p2: Point3f = self.mesh.p[self.mesh.vertex_indices[self.id * 3 + 2]];
+        let p0: Point3f = self.mesh.p[self.mesh.vertex_indices[(self.id * 3) as usize + 0]];
+        let p1: Point3f = self.mesh.p[self.mesh.vertex_indices[(self.id * 3) as usize + 1]];
+        let p2: Point3f = self.mesh.p[self.mesh.vertex_indices[(self.id * 3) as usize + 2]];
         // translate vertices based on ray origin
         let mut p0t: Point3f = p0
             - Vector3f {
@@ -344,16 +344,7 @@ impl Triangle {
         let dndv: Normal3f = Normal3f::default();
         let wo: Vector3f = -ray.d;
         let mut si: SurfaceInteraction = SurfaceInteraction::new(
-            &p_hit,
-            &p_error,
-            &uv_hit,
-            &wo,
-            &dpdu,
-            &dpdv,
-            &dndu,
-            &dndv,
-            ray.time,
-            None,
+            &p_hit, &p_error, &uv_hit, &wo, &dpdu, &dpdv, &dndu, &dndv, ray.time, None,
         );
         // override surface normal in _isect_ for triangle
         let surface_normal: Normal3f = Normal3f::from(vec3_cross_vec3(&dp02, &dp12).normalize());
@@ -365,9 +356,9 @@ impl Triangle {
             // compute shading normal _ns_ for triangle
             let mut ns: Normal3f;
             if !self.mesh.n.is_empty() {
-                let n0 = self.mesh.n[self.mesh.vertex_indices[self.id * 3 + 0]];
-                let n1 = self.mesh.n[self.mesh.vertex_indices[self.id * 3 + 1]];
-                let n2 = self.mesh.n[self.mesh.vertex_indices[self.id * 3 + 2]];
+                let n0 = self.mesh.n[self.mesh.vertex_indices[(self.id * 3) as usize + 0]];
+                let n1 = self.mesh.n[self.mesh.vertex_indices[(self.id * 3) as usize + 1]];
+                let n2 = self.mesh.n[self.mesh.vertex_indices[(self.id * 3) as usize + 2]];
                 ns = Normal3::from(n0) * b0 + Normal3::from(n1) * b1 + Normal3::from(n2) * b2;
                 if ns.length_squared() > 0.0 {
                     ns = ns.normalize();
@@ -380,9 +371,9 @@ impl Triangle {
             // compute shading tangent _ss_ for triangle
             let mut ss: Vector3f;
             if !self.mesh.s.is_empty() {
-                let s0 = self.mesh.s[self.mesh.vertex_indices[self.id * 3 + 0]];
-                let s1 = self.mesh.s[self.mesh.vertex_indices[self.id * 3 + 1]];
-                let s2 = self.mesh.s[self.mesh.vertex_indices[self.id * 3 + 2]];
+                let s0 = self.mesh.s[self.mesh.vertex_indices[(self.id * 3) as usize + 0]];
+                let s1 = self.mesh.s[self.mesh.vertex_indices[(self.id * 3) as usize + 1]];
+                let s2 = self.mesh.s[self.mesh.vertex_indices[(self.id * 3) as usize + 2]];
                 ss = s0 * b0 + s1 * b1 + s2 * b2;
                 if ss.length_squared() > 0.0 {
                     ss = ss.normalize();
@@ -407,12 +398,16 @@ impl Triangle {
                 // compute deltas for triangle partial derivatives of normal
                 let duv02: Vector2f = uv[0] - uv[2];
                 let duv12: Vector2f = uv[1] - uv[2];
-                let dn1: Normal3f =
-                    Normal3::from(self.mesh.n[self.mesh.vertex_indices[self.id * 3 + 0]])
-                        - Normal3::from(self.mesh.n[self.mesh.vertex_indices[self.id * 3 + 2]]);
-                let dn2: Normal3f =
-                    Normal3::from(self.mesh.n[self.mesh.vertex_indices[self.id * 3 + 1]])
-                        - Normal3::from(self.mesh.n[self.mesh.vertex_indices[self.id * 3 + 2]]);
+                let dn1: Normal3f = Normal3::from(
+                    self.mesh.n[self.mesh.vertex_indices[(self.id * 3) as usize + 0]],
+                ) - Normal3::from(
+                    self.mesh.n[self.mesh.vertex_indices[(self.id * 3) as usize + 2]],
+                );
+                let dn2: Normal3f = Normal3::from(
+                    self.mesh.n[self.mesh.vertex_indices[(self.id * 3) as usize + 1]],
+                ) - Normal3::from(
+                    self.mesh.n[self.mesh.vertex_indices[(self.id * 3) as usize + 2]],
+                );
                 let determinant: Float = duv02.x * duv12.y - duv02.y * duv12.x;
                 let degenerate_uv: bool = determinant.abs() < 1e-8;
                 if degenerate_uv {
@@ -442,9 +437,9 @@ impl Triangle {
         // TODO: ProfilePhase p(Prof::TriIntersectP);
         // TODO: ++nTests;
         // get triangle vertices in _p0_, _p1_, and _p2_
-        let p0: Point3f = self.mesh.p[self.mesh.vertex_indices[self.id * 3 + 0]];
-        let p1: Point3f = self.mesh.p[self.mesh.vertex_indices[self.id * 3 + 1]];
-        let p2: Point3f = self.mesh.p[self.mesh.vertex_indices[self.id * 3 + 2]];
+        let p0: Point3f = self.mesh.p[self.mesh.vertex_indices[(self.id * 3) as usize + 0]];
+        let p1: Point3f = self.mesh.p[self.mesh.vertex_indices[(self.id * 3) as usize + 1]];
+        let p2: Point3f = self.mesh.p[self.mesh.vertex_indices[(self.id * 3) as usize + 2]];
         // translate vertices based on ray origin
         let mut p0t: Point3f = p0
             - Vector3f {
@@ -648,17 +643,17 @@ impl Triangle {
     }
     pub fn area(&self) -> Float {
         // get triangle vertices in _p0_, _p1_, and _p2_
-        let p0: Point3f = self.mesh.p[self.mesh.vertex_indices[self.id * 3 + 0]];
-        let p1: Point3f = self.mesh.p[self.mesh.vertex_indices[self.id * 3 + 1]];
-        let p2: Point3f = self.mesh.p[self.mesh.vertex_indices[self.id * 3 + 2]];
+        let p0: Point3f = self.mesh.p[self.mesh.vertex_indices[(self.id * 3) as usize + 0]];
+        let p1: Point3f = self.mesh.p[self.mesh.vertex_indices[(self.id * 3) as usize + 1]];
+        let p2: Point3f = self.mesh.p[self.mesh.vertex_indices[(self.id * 3) as usize + 2]];
         0.5 as Float * vec3_cross_vec3(&(p1 - p0), &(p2 - p0)).length()
     }
     pub fn sample(&self, u: &Point2f, pdf: &mut Float) -> InteractionCommon {
         let b: Point2f = uniform_sample_triangle(u);
         // get triangle vertices in _p0_, _p1_, and _p2_
-        let p0: Point3f = self.mesh.p[self.mesh.vertex_indices[self.id * 3 + 0]];
-        let p1: Point3f = self.mesh.p[self.mesh.vertex_indices[self.id * 3 + 1]];
-        let p2: Point3f = self.mesh.p[self.mesh.vertex_indices[self.id * 3 + 2]];
+        let p0: Point3f = self.mesh.p[self.mesh.vertex_indices[(self.id * 3) as usize + 0]];
+        let p1: Point3f = self.mesh.p[self.mesh.vertex_indices[(self.id * 3) as usize + 1]];
+        let p2: Point3f = self.mesh.p[self.mesh.vertex_indices[(self.id * 3) as usize + 2]];
         let mut it: InteractionCommon = InteractionCommon::default();
         it.p = p0 * b[0] + p1 * b[1] + p2 * (1.0 as Float - b[0] - b[1]);
         // compute surface normal for sampled point on triangle
@@ -667,9 +662,9 @@ impl Triangle {
         // the same approach as was used in Triangle::Intersect().
         if !self.mesh.n.is_empty() {
             let ns: Normal3f = Normal3f::from(
-                self.mesh.n[self.mesh.vertex_indices[self.id * 3 + 0]] * b[0]
-                    + self.mesh.n[self.mesh.vertex_indices[self.id * 3 + 1]] * b[1]
-                    + self.mesh.n[self.mesh.vertex_indices[self.id * 3 + 2]]
+                self.mesh.n[self.mesh.vertex_indices[(self.id * 3) as usize + 0]] * b[0]
+                    + self.mesh.n[self.mesh.vertex_indices[(self.id * 3) as usize + 1]] * b[1]
+                    + self.mesh.n[self.mesh.vertex_indices[(self.id * 3) as usize + 2]]
                         * (1.0 as Float - b[0] - b[1]),
             );
             it.n = nrm_faceforward_nrm(&it.n, &ns);
